Release 1.2.0

+++ src/client/src/components/flow/hooks/useOnRemove.ts
@@ -1,59 +1,68 @@
-import { removeElements, FlowElement, Elements } from "react-flow-renderer";
+import { Elements, FlowElement, removeElements } from "react-flow-renderer";
 import { Dispatch } from "redux";
 import { Size } from "../../../compLibrary/size";
-import { GetSelectedNode, IsAspectNode, IsBlockView, GetSelectedBlockNode, IsOffPage } from "../../../helpers";
-import { EDGE_KIND, Node, Project } from "../../../models";
+import { EDGE_KIND, Edge, Project } from "../../../models";
 import { EDGE_TYPE, MODULE_TYPE } from "../../../models/project";
 import { SetPanelHeight } from "../../../modules/inspector/helpers";
-import { changeInspectorHeight } from "../../../modules/inspector/redux/height/actions";
-import { setModuleVisibility } from "../../../redux/store/modules/actions";
+import { changeInspectorHeight } from "../../../modules/inspector/redux/inspectorSlice";
+import { setModuleVisibility } from "../../../redux/store/modules/modulesSlice";
 import { removeEdge, removeNode, setOffPageStatus } from "../../../redux/store/project/actions";
-import { GetParent, IsTransport } from "../helpers";
+import { HandleOffPageDelete } from "../block/nodes/helpers/offPage";
+import { GetParent, IsPartOf } from "../helpers";
+import { GetSelectedBlockNode, GetSelectedNode, IsAspectNode, IsBlockView, IsOffPage } from "../../../helpers";
+import { getParentNodeConnector } from "../block/nodes/helpers/offPage/HandleOffPageDelete";
 
 const useOnRemove = (
   elements: Elements,
-  setElements: any,
-  dispatch: Dispatch,
+  blockEdgesToRemove: Edge[],
   inspectorRef: React.MutableRefObject<HTMLDivElement>,
-  project: Project
+  project: Project,
+  setElements: React.Dispatch<React.SetStateAction<Elements>>,
+  dispatch: Dispatch
 ) => {
   const elementsToRemove: Elements = [];
-
   elements = elements.filter((el) => !IsAspectNode(el.data));
+  const blockView = IsBlockView();
 
-  const hasDeletedElement = handleDeleteElements(elements, elementsToRemove, project, dispatch);
+  const hasDeletedElement = handleDeleteElements(elements, elementsToRemove, project, blockView, dispatch);
+  if (blockView) handleBlockEdges(blockEdgesToRemove, project, dispatch);
 
   if (hasDeletedElement) {
-    dispatch(setModuleVisibility(MODULE_TYPE.INSPECTOR, false, true));
+    dispatch(setModuleVisibility({ type: MODULE_TYPE.INSPECTOR, visible: false, animate: true }));
     SetPanelHeight(inspectorRef, Size.ModuleClosed);
     dispatch(changeInspectorHeight(Size.ModuleClosed));
     return setElements((els) => removeElements(elementsToRemove, els));
   }
 };
 
-const handleDeleteElements = (elements: Elements, verifiedList: Elements, project: Project, dispatch: Dispatch) => {
+const handleDeleteElements = (
+  elements: Elements,
+  verifiedList: Elements,
+  project: Project,
+  blockView: boolean,
+  dispatch: Dispatch
+) => {
   const selectedNode = GetSelectedNode();
   const selectedBlockNode = GetSelectedBlockNode();
-  const blockView = IsBlockView();
   const edgeTypes = Object.values(EDGE_TYPE);
   let hasDeletedElement = false;
 
-  for (var elem of elements) {
+  for (const elem of elements) {
     const isEdge = isElementEdge(edgeTypes, elem);
 
     if (isEdge) {
-      if (!IsAspectNode(blockView ? selectedBlockNode : selectedNode)) {
-        if (findProjectEdgeByElementId(project, elem)?.isLocked) continue;
+      if (!IsAspectNode(blockView ? selectedBlockNode : selectedNode) && !findProjectEdgeByElementId(project, elem)?.isLocked) {
         hasDeletedElement = true;
+
+        handleRelatedOffPageElements(project, elem?.data?.edge, dispatch);
         dispatch(removeEdge(elem.id));
         verifiedList.push(elem);
       }
     } else {
       const node = findProjectNodeByElementId(project, elem);
-
       if (node?.isLocked) continue;
 
-      if (IsOffPage(node)) handleOffPageDelete(node, project, dispatch);
+      if (IsOffPage(node)) HandleOffPageDelete(project, node, dispatch);
 
       hasDeletedElement = true;
       dispatch(removeNode(elem.id));
@@ -62,20 +71,6 @@ const handleDeleteElements = (elements: Elements, verifiedList: Elements, projec
   }
 
   return hasDeletedElement;
-};
-
-const handleOffPageDelete = (node: Node, project: Project, dispatch: Dispatch) => {
-  const parentNode = GetParent(node);
-
-  const offPageEdge = project.edges.find(
-    (x) =>
-      (x.fromConnector.nodeId === parentNode.id && IsTransport(x.fromConnector) && x.toConnector.nodeId === node.id) ||
-      (x.toConnector.nodeId === parentNode.id && IsTransport(x.toConnector) && x.fromConnector.nodeId === node.id)
-  );
-
-  const parentNodeConnector = offPageEdge.fromConnector.nodeId === node.id ? offPageEdge.toConnector : offPageEdge.fromConnector;
-
-  if (offPageEdge) dispatch(setOffPageStatus(parentNode.id, parentNodeConnector.id, false));
 };
 
 const isElementEdge = (edgeTypes: string[], element: FlowElement) => {
@@ -88,6 +83,44 @@ const findProjectEdgeByElementId = (project: Project, element: FlowElement) => {
 
 const findProjectNodeByElementId = (project: Project, element: FlowElement) => {
   return project.nodes.find((node) => node.id === element.id);
+};
+
+const handleBlockEdges = (edgesToRemove: Edge[], project: Project, dispatch: Dispatch) => {
+  if (edgesToRemove.length !== 0) {
+    edgesToRemove.forEach((edge) => {
+      handleRelatedOffPageElements(project, edge, dispatch);
+      dispatch(removeEdge(edge.id));
+    });
+  }
+};
+
+const handleRelatedOffPageElements = (project: Project, edge: Edge, dispatch: Dispatch) => {
+  project.nodes.forEach((node) => {
+    if (IsOffPage(node) && (node?.id === edge?.fromNodeId || node?.id === edge?.toNodeId)) {
+      const offPageTransportEdge = project.edges.find(
+        (x) =>
+          (IsOffPage(x?.fromNode) || IsOffPage(x?.toNode)) &&
+          (x?.toConnectorId === edge?.toConnectorId || x?.fromConnectorId === edge?.fromConnectorId)
+      );
+
+      if (!offPageTransportEdge) return;
+
+      const offPagePartOfTerminal = node?.connectors?.find((c) => IsPartOf(c));
+
+      const offPagePartOfEdge = project.edges.find(
+        (x) => IsOffPage(x.toNode) && x.toNodeId === node.id && x.toConnectorId === offPagePartOfTerminal?.id
+      );
+
+      const parentNode = GetParent(node);
+      const parentNodeConnector = getParentNodeConnector(offPageTransportEdge, node);
+      const parentConnectorIsRequired = false;
+
+      dispatch(setOffPageStatus(parentNode?.id, parentNodeConnector?.id, parentConnectorIsRequired));
+      dispatch(removeEdge(offPageTransportEdge?.id));
+      dispatch(removeEdge(offPagePartOfEdge?.id));
+      dispatch(removeNode(node.id));
+    }
+  });
 };
 
 export default useOnRemove;
